services/recommender/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0
scikit-learn==1.4.0
numpy==1.26.4
scipy==1.12.0
pytest==8.0.2
httpx==0.26.0
python-multipart==0.0.9
pydantic==2.6.1
joblib==1.3.2 